Consolidate WASM Error Handling
