+++ matchbox_socket/src/webrtc_socket/error.rs
@@ -1,4 +1,5 @@
 use crate::webrtc_socket::messages::PeerEvent;
+use cfg_if::cfg_if;
 use futures_channel::mpsc::TrySendError;
 
 /// An error that can occur with WebRTC signalling.
@@ -26,10 +27,44 @@ pub enum SignallingError {
 }
 
 /// An error that can occur with WebRTC messaging.
-
 #[derive(Debug, thiserror::Error)]
 pub enum MessagingError {
-    // Common
+    // Native
+    #[cfg(not(target_arch = "wasm32"))]
     #[error("failed to send message to peer")]
-    SendError(Option<String>),
+    SendError(#[from] futures_channel::mpsc::TrySendError<crate::Packet>),
+
+    // WASM
+    #[cfg(target_arch = "wasm32")]
+    #[error("failed to send message to peer")]
+    SendError(#[from] JsError),
+}
+
+cfg_if! {
+    if #[cfg(target_arch = "wasm32")] {
+        use wasm_bindgen::{JsValue};
+
+        // The below is just to wrap Result<JsValue, JsValue> into something sensible-ish
+
+        pub trait JsErrorExt<T> {
+            fn efix(self) -> Result<T, JsError>;
+        }
+
+        impl<T> JsErrorExt<T> for Result<T, JsValue> {
+            fn efix(self) -> Result<T, JsError> {
+                self.map_err(JsError)
+            }
+        }
+
+        #[derive(Debug)]
+        pub struct JsError(JsValue);
+
+        impl std::error::Error for JsError {}
+
+        impl std::fmt::Display for JsError {
+            fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+                write!(f, "{:?}", self.0)
+            }
+        }
+    }
 }
